\docType{package}
\name{rcusum}
\alias{rcusum}
\alias{rcusum-package}
\title{rcusum - An Implementation of Cusum Control Charts as an R package}
\description{
  This R Package implements the charts as described by
  _Rogers et al. (2004, doi:10.1016/j.jtcvs.2004.03.011)_.
  Cusum control charts are devices for monitoring arbitrary
  processes for a prespecified event. This package and
  manual uses the term failure instead of event, although
  any other type of event such as successes could be
  monitored as well. The processes of our interest are
  typically surgical procedures.
}
\author{
  Alexander Meyer
}

